backend/app/faq/views.py
from rest_framework import generics, permissions
from .models import FAQ
from .serializers import FAQSerializer

class FAQListView(generics.ListCreateAPIView):
    queryset = FAQ.objects.all()
    serializer_class = FAQSerializer

class FAQDetailView(generics.RetrieveUpdateDestroyAPIView):
    queryset = FAQ.objects.all()
    serializer_class = FAQSerializer
