backend/requirements.txt
Django>=4.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn>=20.1.0
psycopg2-binary>=2.9.1
Pillow>=8.2.0
mysqlclient==2.2.5
bleach==6.1.0
dj-database-url>=1.0
python-dotenv==1.1.1
